Folder structure update: change folder name from community to plogging to better represent the feature
