+++ src/app/plogging/recruit/[slug]/page.tsx
@@ -1,8 +1,8 @@
-import CarouselForThumbnail from '@/features/community/events/CarouselForThumbnail';
-import EventPostContentSection from '@/features/community/main/EventPostContentSection';
-import EventPostNavigationLinks from '@/features/community/main/EventPostNavigationLinks';
-import ShareLinkButtonSet from '@/features/community/events/ShareLinkButtonSet';
-import { getPostDetail } from '@/services/events';
+import CarouselForThumbnail from '@/features/plogging/events/CarouselForThumbnail';
+import EventPostContentSection from '@/features/plogging/main/EventPostContentSection';
+import ImageSlider from '@/features/plogging/main/ImageSlider';
+import ShareLinkButtonSet from '@/features/plogging/events/ShareLinkButtonSet';
+import { getRecruitDetail } from '@/services/recruits';
 import { countDDay } from '@/util/util';
 import Image from 'next/image';
 import React from 'react';
@@ -13,30 +13,27 @@ type Props = {
   };
 };
 
-const CATEGORY_BOX_CLASS = 'mb-3';
-const CATEGORY_CLASS = 'mr-6 inline-block w-[52px] B300 text-gray-07';
-const CONTENT_CLASS = 'B300 text-gray-12';
+const CATEGORY_BOX_CLASS = 'mb-3 flex';
+const CATEGORY_CLASS = 'mr-6 w-[52px] shrink-0 B300 text-gray-07';
+const CONTENT_CLASS = 'B300 text-gray-12 break-words';
 const DIVIDER_CLASS = 'w-full h-2 bg-gray-01';
 
-export default async function EventDetailPage({ params: { slug } }: Props) {
-  const post = await getPostDetail(slug);
+// user 정보 추가, website를 openChatLink로 변경 필요
+export default async function RecruitDetailPage({ params: { slug } }: Props) {
+  const post = await getRecruitDetail(slug);
   const {
     id,
     title,
     beginEvent,
     finishEvent,
-    organizer,
     location,
-    dues,
     venue,
+    images,
     dueDate,
     numOfPeople,
+    website,
     content,
     position,
-    images,
-    website,
-    prev,
-    next,
   } = post;
 
   if (!post) {
@@ -59,7 +56,7 @@ export default async function EventDetailPage({ params: { slug } }: Props) {
     <div className='h-full bg-white'>
       {images?.length > 0 ? (
         <CarouselForThumbnail
-          slides={imageSlider}
+          slides={<ImageSlider images={images} />}
           options={{ loop: false, align: 'start' }}
         />
       ) : (
@@ -67,9 +64,10 @@ export default async function EventDetailPage({ params: { slug } }: Props) {
       )}
       <div className='px-4 pt-3.5 pb-5 text-start'>
         <div className='mb-3.5 flex B100 text-center'>
+          {/* 시안님과 확인 */}
           {countDDay(post.beginEvent) < 0 ? (
             <span className='mr-2 py-0.5 px-1.5 B100 text-gray-07 bg-gray-03 rounded-lg'>
-              행사종료
+              모집마감
             </span>
           ) : (
             <span className='mr-1.5 py-0.5 px-1.5 text-white bg-gray-09 rounded-lg'>
@@ -81,51 +79,37 @@ export default async function EventDetailPage({ params: { slug } }: Props) {
           <span className='mr-1.5 py-0.5 px-1.5 text-gray-12 bg-gray-03 rounded-lg'>
             {post.location}
           </span>
-          {post.dues ? null : (
-            <span className='py-0.5 px-1.5 text-blue-01 bg-blue-03 rounded-lg'>
-              참가비 무료
-            </span>
-          )}
         </div>
         <h1 className='mb-2 H500 text-gray-12'>{title}</h1>
-        <p className='mb-4 B300 text-gray-07'>{organizer}</p>
         <div className='mb-5 w-full h-[1px] bg-gray-02' />
         <div className={CATEGORY_BOX_CLASS}>
-          <span className={CATEGORY_CLASS}>행사 일정</span>
+          <span className={CATEGORY_CLASS}>모임 일정</span>
           <span className={CONTENT_CLASS}>
             {new Date(beginEvent).toLocaleDateString()} -{' '}
-            {new Date(finishEvent).toDateString()}
+            {new Date(finishEvent).toLocaleDateString()}
           </span>
         </div>
         <div className={CATEGORY_BOX_CLASS}>
-          <span className={CATEGORY_CLASS}>행사 장소</span>
+          <span className={CATEGORY_CLASS}>모임 장소</span>
           <span className={CONTENT_CLASS}>{venue}</span>
         </div>
         <div className={CATEGORY_BOX_CLASS}>
           <span className={CATEGORY_CLASS}>모집 기간</span>
           <span className={CONTENT_CLASS}>
-            {new Date(dueDate).toDateString()}까지
+            {new Date(dueDate).toLocaleDateString()}까지
           </span>
         </div>
         <div className={CATEGORY_BOX_CLASS}>
           <span className={CATEGORY_CLASS}>모집 인원</span>
           <span className={CONTENT_CLASS}>{numOfPeople}</span>
         </div>
-        <div className={CATEGORY_BOX_CLASS}>
-          <span className={CATEGORY_CLASS}>참가비</span>
-          {dues ? (
-            <span className={CONTENT_CLASS}>{dues}</span>
-          ) : (
-            <span className='B300 text-blue-01'>무료</span>
-          )}
-        </div>
       </div>
       <div className={DIVIDER_CLASS} />
-      <EventPostNavigationLinks prev={prev} next={next} />
+      {/* user info goes here */}
       <div className={DIVIDER_CLASS} />
       <EventPostContentSection post={post} />
       <div className='w-full'>
-        <ShareLinkButtonSet path={`communtiy/event/${slug}`} url={website} />
+        <ShareLinkButtonSet path={`communtiy/recruit/${slug}`} url={website} />
       </div>
     </div>
   );
