+++ src/features/plogging/main/ImageSlider.tsx
